+++ src/ui/map.test.ts
@@ -20,6 +20,7 @@ import Transform from '../geo/transform';
 import {StyleImageInterface} from '../style/style_image';
 import Style from '../style/style';
 import {MapSourceDataEvent} from './events';
+import config from '../util/config';
 
 function createStyleSource() {
     return {
@@ -164,6 +165,16 @@ describe('Map', () => {
     });
 
     describe('#mapOptions', () => {
+        test('maxTileCacheZoomLevels: Default value is set', () => {
+            const map = createMap();
+            expect(map._maxTileCacheZoomLevels).toBe(config.MAX_TILE_CACHE_ZOOM_LEVELS);
+        });
+
+        test('maxTileCacheZoomLevels: Value can be set via map options', () => {
+            const map = createMap({maxTileCacheZoomLevels: 1});
+            expect(map._maxTileCacheZoomLevels).toBe(1);
+        });
+
         test('Style validation is enabled by default', () => {
             let validationOption = false;
             jest.spyOn(Style.prototype, 'loadJSON').mockImplementationOnce((styleJson, options) => {