+++ src/ui/map.ts
@@ -64,6 +64,7 @@ import type {ControlPosition, IControl} from './control/control';
 import type {MapGeoJSONFeature} from '../util/vectortile_to_geojson';
 import Terrain from '../render/terrain';
 import RenderToTexture from '../render/render_to_texture';
+import config from '../util/config';
 
 const version = packageJSON.version;
 /* eslint-enable no-use-before-define */
@@ -101,6 +102,7 @@ export type MapOptions = {
     pitch?: number;
     renderWorldCopies?: boolean;
     maxTileCacheSize?: number;
+    maxTileCacheZoomLevels?: number;
     transformRequest?: RequestTransformFunction;
     transformCameraUpdate?: CameraUpdateTransformFunction;
     locale?: any;
@@ -194,6 +196,7 @@ const defaultOptions = {
     renderWorldCopies: true,
     refreshExpiredTiles: true,
     maxTileCacheSize: null,
+    maxTileCacheZoomLevels: config.MAX_TILE_CACHE_ZOOM_LEVELS,
     localIdeographFontFamily: 'sans-serif',
     transformRequest: null,
     transformCameraUpdate: null,
@@ -265,7 +268,8 @@ const defaultOptions = {
  * container, there will be blank space beyond 180 and -180 degrees longitude.
  * - Features that cross 180 and -180 degrees longitude will be cut in two (with one portion on the right edge of the
  * map and the other on the left edge of the map) at every zoom level.
- * @param {number} [options.maxTileCacheSize=null] The maximum number of tiles stored in the tile cache for a given source. If omitted, the cache will be dynamically sized based on the current viewport.
+ * @param {number} [options.maxTileCacheSize=null] The maximum number of tiles stored in the tile cache for a given source. If omitted, the cache will be dynamically sized based on the current viewport which can be set using `maxTileCacheZoomLevels` constructor options.
+ * @param {number} [options.maxTileCacheZoomLevels=5] The maximum number of zoom levels for which to store tiles for a given source. Tile cache dynamic size is calculated by multiplying `maxTileCacheZoomLevels` with approx number of tiles in the viewport for a given source.
  * @param {string} [options.validateStyle=true] If false, style validation will be skipped. Useful in production environment.
  * @param {string} [options.localIdeographFontFamily='sans-serif'] Defines a CSS
  * font-family for locally overriding generation of glyphs in the 'CJK Unified Ideographs', 'Hiragana', 'Katakana' and 'Hangul Syllables' ranges.
@@ -321,6 +325,7 @@ class Map extends Camera {
     _vertices: boolean;
     _canvas: HTMLCanvasElement;
     _maxTileCacheSize: number;
+    _maxTileCacheZoomLevels: number;
     _frame: Cancelable;
     _styleDirty: boolean;
     _sourcesDirty: boolean;
@@ -434,6 +439,7 @@ class Map extends Camera {
         this._cooperativeGestures = options.cooperativeGestures;
         this._metaKey = navigator.platform.indexOf('Mac') === 0 ? 'metaKey' : 'ctrlKey';
         this._maxTileCacheSize = options.maxTileCacheSize;
+        this._maxTileCacheZoomLevels = options.maxTileCacheZoomLevels;
         this._failIfMajorPerformanceCaveat = options.failIfMajorPerformanceCaveat;
         this._preserveDrawingBuffer = options.preserveDrawingBuffer;
         this._antialias = options.antialias;