+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ### ✨ Features and improvements
+- Expose map options.maxTileCacheZoomLevels to allow better control of tile cache ([#2581](https://github.com/maplibre/maplibre-gl-js/pull/2581))
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes